Clean-up of the master branch for ansible collection as roles moved to roles branch
