--- action_plugins
@@ -1,1 +1,0 @@
-ansible_collection/Juniper/junos/plugins/action
