--- ansible_collection/Juniper/junos/README.md
@@ -1,3 +1,0 @@
-# Ansible Collection - juniper.junos
-
-Documentation for the collection.
