+++ ansible_collections/Juniper/junos_collection/README.md
@@ -1,0 +1,1 @@
+../../../README.md
