+++ ansible_collections/Juniper/junos_collection/galaxy.yml
@@ -6,7 +6,7 @@
 namespace: Juniper
 
 # The name of the collection. Has the same character restrictions as 'namespace'
-name: junos
+name: junos_collection
 
 # The version of the collection. Must be compatible with semantic versioning
 version: 0.1.0
