+++ ansible_collections/Juniper/junos_collection/plugins/modules/juniper_junos_config.py
@@ -47,7 +47,6 @@ extends_documentation_fragment:
   - juniper_junos_common.connection_documentation
   - juniper_junos_common.logging_documentation
 module: juniper_junos_config
-version_added: "2.0.0" # of Juniper.junos role
 author: "Juniper Networks - Stacy Smith (@stacywsmith)"
 short_description: Manipulate the configuration of a Junos device
 description:
@@ -521,8 +520,8 @@ EXAMPLES = '''
   hosts: junos-all
   connection: local
   gather_facts: no
-  roles:
-    - Juniper.junos
+  collections:
+    - Juniper.junos_collection
   tasks:
     - name: Retrieve the committed configuration
       juniper_junos_config:
@@ -740,7 +739,7 @@ Reference for the issue: https://groups.google.com/forum/#!topic/ansible-project
 
 # Ansiballz packages module_utils into ansible.module_utils
 from ansible.module_utils.basic import AnsibleModule
-from ansible_collections.Juniper.junos.plugins.module_utils import juniper_junos_common
+from ansible_collections.Juniper.junos_collection.plugins.module_utils import juniper_junos_common
 
 def main():
     # Choices which are defined in the common module.
