+++ ansible_collections/Juniper/junos_collection/plugins/modules/juniper_junos_ping.py
@@ -46,7 +46,6 @@ extends_documentation_fragment:
   - juniper_junos_common.connection_documentation
   - juniper_junos_common.logging_documentation
 module: juniper_junos_ping
-version_added: "2.0.0" # of Juniper.junos role
 author: Juniper Networks - Stacy Smith (@stacywsmith)
 short_description: Execute ping from a Junos device
 description:
@@ -146,8 +145,8 @@ EXAMPLES = '''
   hosts: junos-all
   connection: local
   gather_facts: no
-  roles:
-    - Juniper.junos
+  collections:
+    - Juniper.junos_collection
 
   tasks:
     - name: Ping 192.68.1.1 with default parameters. Fails if any packets lost.
@@ -386,7 +385,7 @@ Reference for the issue: https://groups.google.com/forum/#!topic/ansible-project
 
 # Ansiballz packages module_utils into ansible.module_utils
 from ansible.module_utils.basic import AnsibleModule
-from ansible_collections.Juniper.junos.plugins.module_utils import juniper_junos_common
+from ansible_collections.Juniper.junos_collection.plugins.module_utils import juniper_junos_common
 
 def main():
     # The argument spec for the module.
