+++ ansible_collections/Juniper/junos_collection/plugins/modules/juniper_junos_table.py
@@ -48,7 +48,6 @@ extends_documentation_fragment:
   - juniper_junos_common.connection_documentation
   - juniper_junos_common.logging_documentation
 module: juniper_junos_table
-version_added: "2.0.0" # of Juniper.junos role
 author:
   - Jason Edelman (@jedelman8)
   - Updated by Juniper Networks - Stacy Smith (@stacywsmith)
@@ -128,8 +127,8 @@ EXAMPLES = '''
   hosts: junos-all
   connection: local
   gather_facts: no
-  roles:
-    - Juniper.junos
+  collections:
+    - Juniper.junos_collection
 
   tasks:
     - name: Retrieve LLDP Neighbor Information Using PyEZ-included Table
@@ -292,7 +291,7 @@ Reference for the issue: https://groups.google.com/forum/#!topic/ansible-project
 
 # Ansiballz packages module_utils into ansible.module_utils
 from ansible.module_utils.basic import AnsibleModule
-from ansible_collections.Juniper.junos.plugins.module_utils import juniper_junos_common
+from ansible_collections.Juniper.junos_collection.plugins.module_utils import juniper_junos_common
 
 def expand_items(module, data):
     """Recursively expand any table items
