--- library
@@ -1,1 +1,0 @@
-ansible_collection/Juniper/junos/plugins/modules
