--- module_utils
@@ -1,1 +1,0 @@
-ansible_collection/Juniper/junos/plugins/module_utils
