+++ tests/ansible.cfg
@@ -1,4 +1,5 @@
 
 [defaults]
 hash_behaviour=merge
-roles_path = /etc/ansible/roles
+roles_path = ~/.ansible/roles
+
