+++ tests/pb.juniper_junos_facts.yml
@@ -3,15 +3,11 @@
   hosts: all
   connection: local
   gather_facts: no
-  roles:
-    - Juniper.junos
+  collections:
+    - Juniper.junos_collection
   tasks:
     - name: "TEST 1 - Gather Facts"
       juniper_junos_facts:
-        host: "{{ ansible_ssh_host }}"
-        port: "{{ ansible_ssh_port }}"
-        user: "{{ ansible_ssh_user }}"
-        passwd: "{{ ansible_ssh_pass }}"
       ignore_errors: True
       register: test1
