+++ tests/pb.juniper_junos_ping.yml
@@ -3,15 +3,11 @@
   hosts: all
   connection: local
   gather_facts: no
-  roles:
-    - Juniper.junos
+  collections:
+    - Juniper.junos_collection
   tasks:
     - name: "TEST 1 - Ping Google DNS"
       juniper_junos_ping:
-        host: "{{ ansible_ssh_host }}"
-        port: "{{ ansible_ssh_port }}"
-        user: "{{ ansible_ssh_user }}"
-        passwd: "{{ ansible_ssh_pass }}"
         dest_ip: 8.8.8.8
       register: test1
       ignore_errors: True
@@ -26,10 +22,6 @@
 
     - name: "TEST 2 - Ping Wrong IP"
       juniper_junos_ping:
-        host: "{{ ansible_ssh_host }}"
-        port: "{{ ansible_ssh_port }}"
-        user: "{{ ansible_ssh_user }}"
-        passwd: "{{ ansible_ssh_pass }}"
         dest_ip: 8.8.1.1
       register: test2
       ignore_errors: True
@@ -43,10 +35,6 @@
 
     - name: "TEST 3 - Change nbr packets"
       juniper_junos_ping:
-        host: "{{ ansible_ssh_host }}"
-        port: "{{ ansible_ssh_port }}"
-        user: "{{ ansible_ssh_user }}"
-        passwd: "{{ ansible_ssh_pass }}"
         dest_ip: 8.8.8.8
         count: 3
       register: test3
@@ -62,10 +50,6 @@
 
     - name: "TEST 4 - Ping with DF-bit set"
       juniper_junos_ping:
-        host: "{{ ansible_ssh_host }}"
-        port: "{{ ansible_ssh_port }}"
-        user: "{{ ansible_ssh_user }}"
-        passwd: "{{ ansible_ssh_pass }}"
         dest_ip: 8.8.8.8
         count: 3
         do_not_fragment: True
@@ -83,10 +67,6 @@
 
     - name: "TEST 5 - Ping with DF-bit set and size that well exceeds jumbo sizes"
       juniper_junos_ping:
-        host: "{{ ansible_ssh_host }}"
-        port: "{{ ansible_ssh_port }}"
-        user: "{{ ansible_ssh_user }}"
-        passwd: "{{ ansible_ssh_pass }}"
         dest_ip: 8.8.8.8
         count: 3
         do_not_fragment: True
