+++ tests/pb.juniper_junos_pmtud.yml
@@ -3,15 +3,11 @@
   hosts: all
   connection: local
   gather_facts: no
-  roles:
-    - Juniper.junos
+  collections:
+    - Juniper.junos_collection
   tasks:
     - name: "TEST 1 - Check path MTU to Google DNS"
       juniper_junos_pmtud:
-        host: "{{ ansible_ssh_host }}"
-        port: "{{ ansible_ssh_port }}"
-        user: "{{ ansible_ssh_user }}"
-        passwd: "{{ ansible_ssh_pass }}"
         dest_ip: 8.8.8.8
       register: test1
       ignore_errors: True
