--- tests/ravello.ini
@@ -1,18 +1,0 @@
-[all:children]
-junos
-
-[junos]
-vqfx-01
-vqfx-02
-
-###################################
-### Define variables per groups ###
-###################################
-[all:vars]
-ansible_ssh_user=root
-ansible_ssh_pass=Juniper
-ansible_ssh_port=22
-ravello_ci_app_name="Ansible-junos-stdlib Ansible_{{ lookup('env','ANSIBLE_VERSION') }} Travis_{{ lookup('env','TRAVIS_JOB_ID') }} {{ lookup('env','TRAVIS_COMMIT') }}"
-ravello_ci_blueprint="75695295"
-ravello_ci_token="GtHFbCOuKgD1pcfkvCCIgenj6DOtn3VgRLjaYipdideCsiPC1NxJitt1UHfhF0Bf"
-ravello_ci_expiration_time_min=50
